requirements.txt
absl-py==0.12.0
aio-pika==6.8.0
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
alembic==1.4.3
APScheduler==3.7.0
apturl==0.5.2
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
bidict==0.21.2
blinker==1.4
boto3==1.17.51
botocore==1.20.51
Brlapi==0.7.0
bs4==0.0.1
cachetools==4.2.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
cloudpickle==1.6.0
colorama==0.4.3
colorclass==2.2.0
coloredlogs==14.3
colorhash==1.0.3
command-not-found==0.3
configobj==5.0.6
croniter==1.0.12
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-yamlfield==1.0.3
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
fbmessenger==6.0.0
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.11.1
httplib2==0.14.0
httptools==0.1.1
httpx==0.9.3
humanfriendly==9.1
hyperframe==5.2.0
idna==2.8
isodate==0.6.0
jmespath==0.10.0
joblib==0.15.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.3.4
mattermostwrapper==2.2
monotonic==1.5
multidict==4.6.0
mysqlclient==2.0.3
netifaces==0.10.4
networkx==2.5.1
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
paramiko==2.6.0
pexpect==4.6.0
pika==1.2.0
Pillow==7.0.0
prompt-toolkit==2.0.10
protobuf==3.15.8
psutil==5.5.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
pydot==1.4.2
PyGObject==3.36.0
PyJWT==2.0.0
pykwalify==1.8.0
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.17.3
pyTelegramBotAPI==3.7.7
python-apt==2.0.0+ubuntu0.20.4.4
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-editor==1.0.4
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2020.5
pyxdg==0.26
PyYAML==5.3.1
questionary==1.5.2
rasa==2.5.0
rasa-sdk==2.5.0
rasa-x==0.38.1
redis==3.5.3
regex==2020.9.27
reportlab==3.5.34
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rfc3986==1.4.0
rocketchat-API==1.15.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.7
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.6.0
Sanic-Plugins-Framework==0.9.5
scikit-learn==0.24.1
scipy==1.6.2
SecretStorage==2.3.1
sentry-sdk==0.19.5
simplejson==3.16.0
six==1.14.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==4.0.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
sqlparse==0.4.1
systemd-python==234
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.3.2
tensorflow-addons==0.12.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.10.0
tensorflow-probability==0.11.1
tensorflow-text==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
tqdm==4.59.0
twilio==6.50.1
typeguard==2.12.0
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvloop==0.14.0
wadllib==1.3.3
wcwidth==0.2.5
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.5.1
